git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1145 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/rserpool-internals.h
@@ -85,9 +85,9 @@ int rsp_register_tags(int                        sd,
 int rsp_deregister_tags(int            sd,
                        int             flags,
                        struct TagItem* tags);
-int rsp_accept_tags(int                sd,
-                    unsigned long long timeout,
-                    struct TagItem*    tags);
+int rsp_accept_tags(int             sd,
+                    int             timeout,
+                    struct TagItem* tags);
 int rsp_connect_tags(int                  sd,
                      const unsigned char* poolHandle,
                      const size_t         poolHandleSize,
